init: editItemsElement

+++ src/functions/item.js
@@ -1,9 +1,9 @@
 import {itemAPI, getCachedItems, saveItemsToCache, isActive} from "./functions.js"
 
-const creamShopListElement = document.querySelector('.cream-shop-list')
+const creamShopListElement = document.querySelector('.table-list')
 const formElement = document.querySelector('.form')
+const buttonEditElement = document.querySelector('.edit-form__button-edit')
 const editFormElement = document.querySelector('.edit-form')
-const buttonEditElement = document.querySelector('.button-edit')
 
 const cachedItems = getCachedItems()
 let items = cachedItems.length > 0 ? cachedItems : await itemAPI.getItem()
@@ -11,24 +11,23 @@ let items = cachedItems.length > 0 ? cachedItems : await itemAPI.getItem()
 const renderItems = () => {
 
   creamShopListElement.innerHTML = items.reduce((acc, el) => {
-    const {id, title, count, date} = el
-    const item = `
-      <div class="cream-shop-list__item" data-id="${id}">
-        <div class="cream-shop-list__name">
-          <p class="fw-bold">${id}</p>
-          <div class="cream-shop-list__title fw-medium">${title}</div>
-        </div>
-        <div class="cream-shop-list__name">
-          <span class="fw-medium">${count} шт</span>
-          <span class="fw-medium">${date}</span>
-        </div>
-        <div class="button">
-          <button class="button__delete btn btn-danger">Удалить</button>
-          <button class="button__edit btn btn-primary">Редакт</button>
-        </div>
-      </div>
-      `
+    const {id, category_id, name, manufacturer, price} = el
 
+    const item = `
+    <tr class="table-list__item" data-id="${id}">
+      <th class="table-list__id">${id}</th>
+      <td class="table-list__name">${name}</td>
+      <td class="table-list__manufacturer">${manufacturer}</td>
+      <td class="table-list__category"></td>
+      <td class="table-list__price">${price}</td>
+      <td>
+        <button class="button__delete btn btn-danger">Удалить</button>
+      </td>
+      <td>
+        <button class="button__edit btn btn-primary">Редакт</button>
+     </td>
+      </tr>
+      `
     return acc + item
   }, '')
 }
@@ -37,7 +36,7 @@ saveItemsToCache(items)
 renderItems()
 
 const removeItemsFromDom = (id) => {
-  const itemsElement = document.querySelector(`.cream-shop-list__item[data-id="${id}"]`)
+  const itemsElement = document.querySelector(`.table-list__item[data-id="${id}"]`)
 
   itemsElement?.remove()
 }
@@ -54,13 +53,31 @@ const deleteItems = (id) => {
     })
 }
 
+const editItemsElement = (id) => {
+  const itemsElement = document.querySelector(`.table-list__item[data-id="${id}"]`)
+  const name = itemsElement.querySelector('.table-list__name').textContent
+  const manufacturer = itemsElement.querySelector('.table-list__manufacturer').textContent
+  const category = itemsElement.querySelector('.table-list__category').textContent
+  const price = itemsElement.querySelector('.table-list__price').textContent
+  isActive(id, category, name, manufacturer, price)
+}
+
 const onItemsListClick = (event) => {
   const itemDeleteElement = event.target.closest('.button__delete')
+  const itemEditElement = event.target.closest('.button__edit')
 
-  const itemElement = itemDeleteElement.closest('.cream-shop-list__item')
-  const {id} = itemElement.dataset
+  if (itemDeleteElement) {
+    const itemElement = itemDeleteElement.closest('.table-list__item')
+    const {id} = itemElement.dataset
 
-  deleteItems(id)
+    return deleteItems(id)
+  }
+  if (itemEditElement) {
+    const itemElement = itemEditElement.closest('.table-list__item')
+    const {id} = itemElement.dataset
+
+    editItemsElement(id)
+  }
 }
 
 creamShopListElement.addEventListener('click', onItemsListClick)
@@ -68,9 +85,9 @@ creamShopListElement.addEventListener('click', onItemsListClick)
 const onItemsAddFormSubmit = (event) => {
   event.preventDefault()
   const formDataElement = new FormData(event.target)
-  const {title, count, date} = Object.fromEntries(formDataElement)
-  const id =  items.length += 1
-  const newItemsElement = {id, title, count, date}
+  const {name, manufacturer, price} = Object.fromEntries(formDataElement)
+  const id = items.length + 1
+  const newItemsElement = {id, name, manufacturer, price}
 
   itemAPI.addItem(newItemsElement)
     .then(() => {
@@ -84,13 +101,30 @@ const onItemsAddFormSubmit = (event) => {
 
 }
 
+const onItemsEditFormSubmit = (event) => {
+  event.preventDefault()
+  const formDataElement = new FormData(editFormElement)
+  const {id, name, manufacturer, price} = Object.fromEntries(formDataElement)
+  itemAPI.editItems(id, name, manufacturer, price)
+    .then(() => {
+      const editedItemIndex = items.findIndex(item => item.id === String(id))
+      if (editedItemIndex !== -1) {
+        items[editedItemIndex] = {id, name, manufacturer, price}
+        saveItemsToCache(items)
+        renderItems()
+        console.log(items)
+      }
+    })
+    .catch(err => {
+      console.log(err)
+    })
+
+  editFormElement.classList.remove('is-active')
+}
+
+
 formElement.addEventListener('submit', onItemsAddFormSubmit)
-// buttonEditElement.addEventListener('click', isActive)
-
-
-
-
-
+buttonEditElement.addEventListener('click', onItemsEditFormSubmit)
